Moved InputModule Cursor scripts to UX folder and updated namespace
