+++ Assets/MixedRealityToolkit/UX/Scripts/Cursors/AnimatedCursor.cs.meta
