+++ Assets/MixedRealityToolkit/UX/Scripts/Cursors/AnimatorParameter.cs
@@ -4,7 +4,7 @@
 using System;
 using UnityEngine;
 
-namespace MixedRealityToolkit.InputModule.Cursor
+namespace MixedRealityToolkit.UX.Cursor
 {
     /// <summary>
     /// Created a copy of the AnimatorControllerParameter because that class is not Serializable
