+++ Assets/MixedRealityToolkit/UX/Scripts/Cursors/CursorModifier.cs.meta
