+++ Assets/MixedRealityToolkit/UX/Scripts/Cursors/ICursorModifier.cs
@@ -3,7 +3,7 @@
 
 using UnityEngine;
 
-namespace MixedRealityToolkit.InputModule.Cursor
+namespace MixedRealityToolkit.UX.Cursor
 {
     /// <summary>
     /// Cursor Modifier Interface that provides basic overrides for cursor behavior.
