+++ Assets/MixedRealityToolkit/UX/Scripts/Cursors/ObjectCursor.cs
@@ -4,45 +4,44 @@
 using System;
 using UnityEngine;
 
-namespace MixedRealityToolkit.InputModule.Cursor
+namespace MixedRealityToolkit.UX.Cursor
 {
     /// <summary>
-    /// Object that represents a cursor comprised of sprites and colors for each state
+    /// The object cursor can switch between different game objects based on its state.
+    /// It simply links the game object to set to active with its associated cursor state.
     /// </summary>
-    public class SpriteCursor : BaseCursor
+    public class ObjectCursor : BaseCursor
     {
         [Serializable]
-        public struct SpriteCursorDatum
+        public struct ObjectCursorDatum
         {
             public string Name;
             public CursorStateEnum CursorState;
-            public Sprite CursorSprite;
-            public Color CursorColor;
+            public GameObject CursorObject;
         }
 
         [SerializeField]
-        public SpriteCursorDatum[] CursorStateData;
+        public ObjectCursorDatum[] CursorStateData;
 
         /// <summary>
         /// Sprite renderer to change.  If null find one in children
         /// </summary>
-        public SpriteRenderer TargetRenderer;
+        public Transform ParentTransform;
 
         /// <summary>
         /// On enable look for a sprite renderer on children
         /// </summary>
         protected override void OnEnable()
         {
-            if(TargetRenderer == null)
+            if(ParentTransform == null)
             {
-                TargetRenderer = GetComponentInChildren<SpriteRenderer>();
+                ParentTransform = transform;
             }
-
             base.OnEnable();
         }
 
         /// <summary>
-        /// Override OnCursorState change to set the correct sprite
+        /// Override OnCursorState change to set the correct animation
         /// state for the cursor
         /// </summary>
         /// <param name="state"></param>
@@ -52,30 +51,21 @@ namespace MixedRealityToolkit.InputModule.Cursor
 
             if (state != CursorStateEnum.Contextual)
             {
+                // Hide all children first
+                for(int i = 0; i < ParentTransform.childCount; i++)
+                {
+                    ParentTransform.GetChild(i).gameObject.SetActive(false);
+                }
+
+                // Set active any that match the current state
                 for (int i = 0; i < CursorStateData.Length; i++)
                 {
                     if (CursorStateData[i].CursorState == state)
                     {
-                        SetCursorState(CursorStateData[i]);
+                        CursorStateData[i].CursorObject.SetActive(true);
                     }
                 }
             }
         }
-
-        /// <summary>
-        /// Based on the type of state info pass it through to the sprite renderer
-        /// </summary>
-        /// <param name="stateDatum"></param>
-        private void SetCursorState(SpriteCursorDatum stateDatum)
-        {
-            // Return if we do not have an animator
-            if (TargetRenderer != null)
-            {
-                TargetRenderer.sprite = stateDatum.CursorSprite;
-                TargetRenderer.color = stateDatum.CursorColor;
-            }
-        }
-
     }
-
 }
