+++ Assets/MixedRealityToolkit/UX/Scripts/Cursors/SpriteCursor.cs.meta
